+++ packages/common/src/sites/HubSites.ts
@@ -84,6 +84,7 @@ const DEFAULT_SITE: Partial<IHubSite> = {
       feedback: true,
     },
   },
+  telemetry: {},
 };
 
 /**